141.Last-touches
Categories list moved to CategoriesList  to avoid props-drilling in Home
