+++ crwn-clothing/src/components/directory-item/directory-item.component.jsx
@@ -1,9 +1,11 @@
 import {BackgroundImage, Body, DirectoryItemContainer} from'./directory-item.sytyles'
-
+import { useNavigate } from 'react-router-dom';
 const DirectoryItem = ({ category }) => {
-    const { imageUrl, title } = category;
+    const { imageUrl, title, route } = category;
+    const navigate = useNavigate();
+    const onNavigateHandler = () => navigate(route);
     return (
-      <DirectoryItemContainer>
+      <DirectoryItemContainer onclick={onNavigateHandler}>
         <BackgroundImage imageUrl={imageUrl}
         />
         <Body>
